server/requirements.txt
gunicorn==20.0.4
flask==1.1.2
Flask-Cors==3.0.10
psycopg2-binary==2.8.6
pytest==6.2.1
pytest-mock==3.5.0
pandas==1.2.0